requirements.txt
APScheduler==3.9.1.post1
asgiref==3.6.0
Django==4.1.3
djangorestframework==3.14.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==4.3
